feat: handleClickBackground -> useModal로 빼기 (#17)
