+++ src/hooks/useModal.ts
@@ -1,9 +1,11 @@
 import { useRecoilState } from "recoil";
 
 import { modalState } from "atoms";
+import { useRef } from "react";
 
 const useModal = () => {
   const [modalDataState, setModalDataState] = useRecoilState(modalState);
+  const backgroundRef = useRef<HTMLDivElement>(null);
 
   const handleOpenModal = (modal: React.ReactNode) => {
     const newModal = [...modalDataState];
@@ -17,10 +19,20 @@ const useModal = () => {
     setModalDataState(newModal);
   };
 
+  const handleClickBackground = (e: React.MouseEvent<HTMLDivElement>) => {
+    if (e.target === backgroundRef.current) {
+      handleCloseModal();
+    } else if (backgroundRef.current?.querySelector("#modal-container")) {
+      return null;
+    }
+  };
+
   return {
     modalDataState,
     handleOpenModal,
     handleCloseModal,
+    handleClickBackground,
+    backgroundRef,
   };
 };
